+++ src/Centreon/Tests/Application/Validation/Validator/RepositoryCallbackValidatorTest.php
@@ -40,8 +40,8 @@ use PHPUnit\Framework\TestCase;
 use Centreon\Application\Validation\Validator\RepositoryCallbackValidator;
 use Centreon\Application\Validation\Constraints\RepositoryCallback;
 use Centreon\ServiceProvider;
-use Centreon\Tests\Resource\Dependency;
-use Centreon\Tests\Resource\Mock;
+use Centreon\Tests\Resources\Dependency;
+use Centreon\Tests\Resources\Mock;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\Context\ExecutionContext;
 use Symfony\Component\Validator\Violation\ConstraintViolationBuilder;
