+++ src/Centreon/Tests/Application/Webservice/CentreonFrontendComponentTest.php
@@ -41,7 +41,7 @@ use Pimple\Container;
 use Centreon\ServiceProvider;
 use Centreon\Application\Webservice\CentreonFrontendComponent;
 use Centreon\Domain\Service\FrontendComponentService;
-use Centreon\Tests\Resource\Traits;
+use Centreon\Tests\Resources\Traits;
 
 /**
  * @group Centreon
