+++ src/Centreon/Tests/Infrastructure/CentreonLegacyDB/Mapping/ClassMetadataTest.php
@@ -38,7 +38,7 @@ namespace Centreon\Tests\Infrastructure\CentreonLegacyDB\Mapping;
 
 use PHPUnit\Framework\TestCase;
 use Centreon\Infrastructure\CentreonLegacyDB\Mapping\ClassMetadata;
-use Centreon\Tests\Resource\Mock\EntityMock;
+use Centreon\Tests\Resources\Mock\EntityMock;
 use PDO;
 
 class ClassMetadataTest extends TestCase
