+++ src/Centreon/Tests/Infrastructure/CentreonLegacyDB/ServiceEntityRepositoryTest.php
@@ -42,8 +42,8 @@ use Centreon\Infrastructure\CentreonLegacyDB\EntityPersister;
 use Centreon\Infrastructure\CentreonLegacyDB\Mapping\ClassMetadata;
 use Centreon\Test\Mock\CentreonDB;
 use Centreon\Test\Mock\CentreonDBManagerService;
-use Centreon\Tests\Resource\Mock;
-use Centreon\Tests\Resource\CheckPoint;
+use Centreon\Tests\Resources\Mock;
+use Centreon\Tests\Resources\CheckPoint;
 use Centreon\Test\Traits\TestCaseExtensionTrait;
 
 class ServiceEntityRepositoryTest extends TestCase
