+++ src/Centreon/Tests/Infrastructure/Service/CentreonWebserviceServiceTest.php
@@ -40,7 +40,7 @@ namespace Centreon\Tests\Infrastructure\Service;
 use PHPUnit\Framework\TestCase;
 use Psr\Container\ContainerInterface;
 use Centreon\Infrastructure\Service\CentreonWebserviceService;
-use Centreon\Tests\Resource\Mock\WebserviceMock;
+use Centreon\Tests\Resources\Mock\WebserviceMock;
 use Centreon\Infrastructure\Service\Exception\NotFoundException;
 
 class CentreonWebserviceServiceTest extends TestCase
