+++ src/Centreon/Tests/Resources/Fixture/AclGroup/response-list-1.json
