+++ src/Centreon/Tests/Resources/Fixture/AclGroup/response-list-2.json
