+++ src/Centreon/Tests/Resources/Fixture/ContactGroups/response-list-1.json
