+++ src/Centreon/Tests/Resources/Fixture/ContactGroups/response-list-2.json
