+++ src/Centreon/Tests/Resources/Fixture/Images/response-list-1.json
