+++ src/Centreon/Tests/Resources/Fixture/Images/response-list-2.json
