+++ src/Centreon/Tests/Resources/Fixture/Topology/navigation-list-01.json
