+++ src/CentreonCommand/Tests/ServiceProviderTest.php
@@ -24,7 +24,7 @@ namespace CentreonCommand\Tests;
 use PHPUnit\Framework\TestCase;
 use Pimple\Container;
 use CentreonCommand\ServiceProvider;
-use Centreon\Tests\Resource\Traits\WebserviceTrait;
+use Centreon\Tests\Resources\Traits\WebserviceTrait;
 use CentreonCommand\Application\Webservice;
 
 /**
