+++ src/CentreonNotification/Tests/Domain/Repository/EscalationRepositoryTest.php
@@ -25,7 +25,7 @@ use PHPUnit\Framework\TestCase;
 use Centreon\Test\Mock\CentreonDB;
 use CentreonNotification\Domain\Entity\Escalation;
 use CentreonNotification\Domain\Repository\EscalationRepository;
-use Centreon\Tests\Resource\Traits;
+use Centreon\Tests\Resources\Traits;
 
 /**
  * @group CentreonNotification
