+++ src/CentreonUser/Tests/Domain/Repository/ContactRepositoryTest.php
@@ -24,7 +24,7 @@ namespace CentreonUser\Tests\Domain\Repository;
 use CentreonUser\Domain\Repository\ContactRepository;
 use CentreonUser\Domain\Entity\Contact;
 use PHPUnit\Framework\TestCase;
-use Centreon\Tests\Resource\Traits;
+use Centreon\Tests\Resources\Traits;
 
 /**
  * @group CentreonUser
