+++ src/CentreonUser/Tests/ServiceProviderTest.php
@@ -24,7 +24,7 @@ namespace CentreonUser\Tests;
 use PHPUnit\Framework\TestCase;
 use Pimple\Container;
 use CentreonUser\ServiceProvider;
-use Centreon\Tests\Resource\Traits\WebserviceTrait;
+use Centreon\Tests\Resources\Traits\WebserviceTrait;
 use CentreonUser\Application\Webservice;
 
 /**
